feat: added functionality for retrieval of business partners with attributes

+++ TotalOne/TotalOne.Api/Requests/FilterBusinessPartnersRequest.cs
@@ -2,7 +2,7 @@
 {
     public class FilterBusinessPartnersRequest
     {
-        public int BusinessPartnerId { get; set; }
+        public long BusinessPartnerId { get; set; }
         public DateTime? LastUpdateStart { get; set; }
         public DateTime? LastUpdateEnd { get; set; }
         public string? Name { get; set; }
